Remove redundant parser code
