+++ src/som/compiler/ast/parser.py
@@ -1,7 +1,8 @@
-from rpython.rlib.rbigint import rbigint
-from rpython.rlib.rstring import ParseStringOverflowError
 from rtruffle.source_section import SourceSection
-from ..parse_error import ParseError, ParseErrorSymList
+
+from .method_generation_context import MethodGenerationContext
+from ..parse_error import ParseError
+from ..parser import ParserBase
 
 from ...interpreter.ast.nodes.block_node import BlockNode, BlockNodeWithContext
 from ...interpreter.ast.nodes.global_read_node import UninitializedGlobalReadNode
@@ -10,151 +11,14 @@ from ...interpreter.ast.nodes.message.uninitialized_node import UninitializedMes
 from ...interpreter.ast.nodes.return_non_local_node import ReturnNonLocalNode
 from ...interpreter.ast.nodes.sequence_node import SequenceNode
 
-from ..lexer import Lexer
 from ..symbol import Symbol
 
-from .method_generation_context import MethodGenerationContext
 
-
-class Parser(object):
-
-    _single_op_syms        = [Symbol.Not,  Symbol.And,  Symbol.Or,    Symbol.Star,
-                              Symbol.Div,  Symbol.Mod,  Symbol.Plus,  Symbol.Equal,
-                              Symbol.More, Symbol.Less, Symbol.Comma, Symbol.At,
-                              Symbol.Per,  Symbol.NONE]
-
-    _binary_op_syms        = [Symbol.Or,   Symbol.Comma, Symbol.Minus, Symbol.Equal,
-                              Symbol.Not,  Symbol.And,   Symbol.Or,    Symbol.Star,
-                              Symbol.Div,  Symbol.Mod,   Symbol.Plus,  Symbol.Equal,
-                              Symbol.More, Symbol.Less,  Symbol.Comma, Symbol.At,
-                              Symbol.Per,  Symbol.NONE]
-
-    _keyword_selector_syms = [Symbol.Keyword, Symbol.KeywordSequence]
+class Parser(ParserBase):
 
     def __init__(self, reader, file_name, universe):
-        self._universe = universe
+        ParserBase.__init__(self, reader, file_name, universe)
         self._source_reader = reader
-        self._file_name = file_name
-        self._lexer    = Lexer(reader)
-        self._sym      = Symbol.NONE
-        self._text     = None
-        self._next_sym = Symbol.NONE
-        self._get_symbol_from_lexer()
-
-    def classdef(self, cgenc):
-        cgenc.set_name(self._universe.symbol_for(self._text))
-        self._expect(Symbol.Identifier)
-        self._expect(Symbol.Equal)
-
-        self._superclass(cgenc)
-
-        self._expect(Symbol.NewTerm)
-        self._instance_fields(cgenc)
-
-        while (self._sym_is_identifier() or self._sym == Symbol.Keyword or
-               self._sym == Symbol.OperatorSequence or
-               self._sym_in(self._binary_op_syms)):
-            mgenc = MethodGenerationContext(self._universe)
-            mgenc.set_holder(cgenc)
-            mgenc.add_argument("self")
-
-            method_body = self._method(mgenc)
-            cgenc.add_instance_method(mgenc.assemble(method_body))
-
-        if self._accept(Symbol.Separator):
-            cgenc.set_class_side(True)
-            self._class_fields(cgenc)
-
-            while (self._sym_is_identifier()      or
-                   self._sym == Symbol.Keyword    or
-                   self._sym == Symbol.OperatorSequence or
-                   self._sym_in(self._binary_op_syms)):
-                mgenc = MethodGenerationContext(self._universe)
-                mgenc.set_holder(cgenc)
-                mgenc.add_argument("self")
-
-                method_body = self._method(mgenc)
-                cgenc.add_class_method(mgenc.assemble(method_body))
-
-        self._expect(Symbol.EndTerm)
-
-    def _superclass(self, cgenc):
-        if self._sym == Symbol.Identifier:
-            super_name = self._universe.symbol_for(self._text)
-            self._accept(Symbol.Identifier)
-        else:
-            super_name = self._universe.symbol_for("Object")
-
-        cgenc.set_super_name(super_name)
-
-        # Load the super class, if it is not nil (break the dependency cycle)
-        if super_name.get_embedded_string() != "nil":
-            super_class = self._universe.load_class(super_name)
-            if not super_class:
-                raise ParseError("Super class %s could not be loaded"
-                                 % super_name.get_embedded_string(), Symbol.NONE, self)
-            cgenc.set_instance_fields_of_super(
-                super_class.get_instance_fields())
-            cgenc.set_class_fields_of_super(
-                super_class.get_class(self._universe).get_instance_fields())
-        else:
-            # TODO: figure out what this is
-            #raise Exception("What is going on here, not in Java, and I don't think we still got a 'class' field")
-            # WARNING:
-            # We hardcode here the field names for Class
-            # since Object class superclass = Class
-            # We avoid here any kind of dynamic solution to avoid further
-            # complexity. However, that makes it static, it is going to make it
-            #  harder to change the definition of Class and Object
-            field_names_of_class = ["class", "superClass", "name",
-                                    "instanceFields", "instanceInvokables"]
-            field_names = self._universe.new_array_with_strings(field_names_of_class)
-            cgenc.set_class_fields_of_super(field_names)
-
-    def _sym_in(self, symbol_list):
-        return self._sym in symbol_list
-
-    def _sym_is_identifier(self):
-        return self._sym == Symbol.Identifier or self._sym == Symbol.Primitive
-
-    def _accept(self, s):
-        if self._sym == s:
-            self._get_symbol_from_lexer()
-            return True
-        return False
-
-    def _accept_one_of(self, symbol_list):
-        if self._sym_in(symbol_list):
-            self._get_symbol_from_lexer()
-            return True
-        return False
-
-    def _expect(self, s):
-        if self._accept(s):
-            return True
-        raise ParseError("Unexpected symbol. Expected %(expected)s, but found "
-                         "%(found)s", s, self)
-
-    def _expect_one_of(self, symbol_list):
-        if self._accept_one_of(symbol_list):
-            return True
-        raise ParseErrorSymList("Unexpected symbol. Expected one of "
-                                "%(expected)s, but found %(found)s",
-                                symbol_list, self)
-
-    def _instance_fields(self, cgenc):
-        if self._accept(Symbol.Or):
-            while self._sym_is_identifier():
-                var = self._variable()
-                cgenc.add_instance_field(self._universe.symbol_for(var))
-            self._expect(Symbol.Or)
-
-    def _class_fields(self, cgenc):
-        if self._accept(Symbol.Or):
-            while self._sym_is_identifier():
-                var = self._variable()
-                cgenc.add_class_field(self._universe.symbol_for(var))
-            self._expect(Symbol.Or)
 
     def _get_source_section(self, coord):
         return SourceSection(
@@ -166,80 +30,11 @@ class Parser(object):
         node.assign_source_section(self._get_source_section(coord))
         return node
 
-    def _method(self, mgenc):
-        self._pattern(mgenc)
-        self._expect(Symbol.Equal)
-        if self._sym == Symbol.Primitive:
-            mgenc.set_primitive()
-            return self._primitive_block()
-        else:
-            return self._method_block(mgenc)
-
-    def _primitive_block(self):
-        self._expect(Symbol.Primitive)
-        return None
-
-    def _pattern(self, mgenc):
-        if self._sym_is_identifier():
-            self._unary_pattern(mgenc)
-        elif self._sym == Symbol.Keyword:
-            self._keyword_pattern(mgenc)
-        else:
-            self._binary_pattern(mgenc)
-
-    def _unary_pattern(self, mgenc):
-        mgenc.set_signature(self._unary_selector())
-
-    def _binary_pattern(self, mgenc):
-        mgenc.set_signature(self._binary_selector())
-        mgenc.add_argument_if_absent(self._argument())
-
-    def _keyword_pattern(self, mgenc):
-        kw = self._keyword()
-        mgenc.add_argument_if_absent(self._argument())
-
-        while self._sym == Symbol.Keyword:
-            kw += self._keyword()
-            mgenc.add_argument_if_absent(self._argument())
-
-        mgenc.set_signature(self._universe.symbol_for(kw))
-
     def _method_block(self, mgenc):
         self._expect(Symbol.NewTerm)
         method_body = self._block_contents(mgenc)
         self._expect(Symbol.EndTerm)
         return method_body
-
-    def _unary_selector(self):
-        return self._universe.symbol_for(self._identifier())
-
-    def _binary_selector(self):
-        s = self._text
-
-        if    self._accept(Symbol.Or):                     pass
-        elif  self._accept(Symbol.Comma):                  pass
-        elif  self._accept(Symbol.Minus):                  pass
-        elif  self._accept(Symbol.Equal):                  pass
-        elif  self._accept_one_of(self._single_op_syms):   pass
-        elif  self._accept(Symbol.OperatorSequence):       pass
-        else: self._expect(Symbol.NONE)
-
-        return self._universe.symbol_for(s)
-
-    def _identifier(self):
-        s = self._text
-        is_primitive = self._accept(Symbol.Primitive)
-        if not is_primitive:
-            self._expect(Symbol.Identifier)
-        return s
-
-    def _keyword(self):
-        s = self._text
-        self._expect(Symbol.Keyword)
-        return s
-
-    def _argument(self):
-        return self._variable()
 
     def _block_contents(self, mgenc):
         if self._accept(Symbol.Or):
@@ -247,10 +42,6 @@ class Parser(object):
             self._expect(Symbol.Or)
 
         return self._block_body(mgenc)
-
-    def _locals(self, mgenc):
-        while self._sym_is_identifier():
-            mgenc.add_local_if_absent(self._variable())
 
     def _block_body(self, mgenc):
         coordinate = self._lexer.get_source_coordinate()
@@ -297,14 +88,6 @@ class Parser(object):
             return self._assign_source(node, coord)
         else:
             return exp
-
-    def _expression(self, mgenc):
-        self._peek_for_next_symbol_from_lexer()
-
-        if self._next_sym == Symbol.Assign:
-            return self._assignation(mgenc)
-        else:
-            return self._evaluation(mgenc)
 
     def _assignation(self, mgenc):
         return self._assignments(mgenc)
@@ -371,9 +154,6 @@ class Parser(object):
             return self._assign_source(result, coordinate)
 
         return self._literal()
-
-    def _variable(self):
-        return self._identifier()
 
     def _messages(self, mgenc, receiver):
         msg = receiver
@@ -478,51 +258,6 @@ class Parser(object):
         else:
             return self._literal_decimal(False)
 
-    def _literal_decimal(self, negate_value):
-        if self._sym == Symbol.Integer:
-            return self._literal_integer(negate_value)
-        else:
-            if self._sym == Symbol.Double:
-                return self._literal_double(negate_value)
-            else:
-                raise ParseError("Could not parse double. "
-                                 "Expected a number but got '%s'" % self._text,
-                                 Symbol.Double, self)
-
-    def _negative_decimal(self):
-        self._expect(Symbol.Minus)
-        return self._literal_decimal(True)
-
-    def _literal_integer(self, negate_value):
-        try:
-            i = int(self._text)
-            if negate_value:
-                i = 0 - i
-            result = self._universe.new_integer(i)
-        except ParseStringOverflowError:
-            bigint = rbigint.fromstr(self._text)
-            if negate_value:
-                bigint.sign = -1
-            result = self._universe.new_biginteger(bigint)
-        except ValueError:
-            raise ParseError("Could not parse integer. "
-                             "Expected a number but got '%s'" % self._text,
-                             Symbol.NONE, self)
-        self._expect(Symbol.Integer)
-        return result
-
-    def _literal_double(self, negate_value):
-        try:
-            f = float(self._text)
-            if negate_value:
-                f = 0.0 - f
-        except ValueError:
-            raise ParseError("Could not parse double. "
-                             "Expected a number but got '%s'" % self._text,
-                             Symbol.NONE, self)
-        self._expect(Symbol.Double)
-        return self._universe.new_double(f)
-
     def _literal_symbol(self):
         self._expect(Symbol.Pound)
         if self._sym == Symbol.STString:
@@ -561,26 +296,6 @@ class Parser(object):
             raise ParseError("Could not parse literal array value",
                              Symbol.NONE, self)
 
-    def _selector(self):
-        if (self._sym == Symbol.OperatorSequence or
-            self._sym_in(self._single_op_syms)):
-            return self._binary_selector()
-        if (self._sym == Symbol.Keyword or
-            self._sym == Symbol.KeywordSequence):
-            return self._keyword_selector()
-        return self._unary_selector()
-
-    def _keyword_selector(self):
-        s = self._text
-        self._expect_one_of(self._keyword_selector_syms)
-        symb = self._universe.symbol_for(s)
-        return symb
-
-    def _string(self):
-        s = self._text
-        self._expect(Symbol.STString)
-        return s
-
     def _nested_block(self, mgenc):
         self._expect(Symbol.NewBlock)
 
@@ -601,18 +316,6 @@ class Parser(object):
         expressions = self._block_contents(mgenc)
         self._expect(Symbol.EndBlock)
         return expressions
-
-    def _block_pattern(self, mgenc):
-        self._block_arguments(mgenc)
-        self._expect(Symbol.Or)
-
-    def _block_arguments(self, mgenc):
-        self._expect(Symbol.Colon)
-        mgenc.add_argument_if_absent(self._argument())
-
-        while self._sym == Symbol.Colon:
-            self._accept(Symbol.Colon)
-            mgenc.add_argument_if_absent(self._argument())
 
     def _variable_read(self, mgenc, variable_name):
         # 'super' needs to be handled separately
@@ -660,14 +363,4 @@ class Parser(object):
             raise RuntimeError("Neither a variable nor a field found in current"
                                " scope that is named " + variable_name + ".")
 
-    def _get_symbol_from_lexer(self):
-        self._sym  = self._lexer.get_sym()
-        self._text = self._lexer.get_text()
-
-    def _peek_for_next_symbol_from_lexer_if_necessary(self):
-        if not self._lexer.get_peek_done():
-            self._peek_for_next_symbol_from_lexer()
-
-    def _peek_for_next_symbol_from_lexer(self):
-        self._next_sym = self._lexer.peek()
